Add ignoreUnkonwFieldsInExportedMetadata option
Setting this option to true will ignore unknown fields in exported metadata.
This can be helpful, because the official schema definitions are not
up-to-date, and could be fields added that are not present in the schem and so
we cannot handle them. The generic workaround is to ignore these fields in
general with the hasuraconf.ignoreUnkonwFieldsInExportedMetadata property or
setting the hasuraConfiguratior.ignoreUnkonwFieldsInExportedMetadata to true.